requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai==1.51.0
python-multipart==0.0.12
websockets==13.1
pytest==8.3.3
httpx==0.27.2
python-dotenv
requests
tenacity
pyyaml>=6.0.1
gradio>=4.0.0